test: 테스트 컨테이너 1개로 줄이기
