+++ whispia-api/src/test/kotlin/com/whispia/api/worry/domain/WorryTest.kt
@@ -1,5 +1,6 @@
 package com.whispia.api.worry.domain
 
+import com.whispia.api.global.BaseTestContainer
 import com.whispia.api.user.domain.User
 import com.whispia.api.user.domain.UserStatus
 import org.assertj.core.api.Assertions.assertThat
@@ -7,44 +8,21 @@ import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 import org.springframework.test.context.ActiveProfiles
-import org.springframework.test.context.DynamicPropertyRegistry
-import org.springframework.test.context.DynamicPropertySource
 import org.springframework.test.context.TestConstructor
 import org.springframework.transaction.annotation.Transactional
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.junit.jupiter.Container
 import org.testcontainers.junit.jupiter.Testcontainers
 import java.util.UUID
 
 @DataJpaTest
 @TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
-@EnableJpaAuditing
 @Transactional
 @Testcontainers
 @ActiveProfiles("test")
 @DisplayName("Worry JPA 테스트")
 class WorryTest(
     private val entityManager: TestEntityManager
-) {
-
-    companion object {
-        @Container
-        @JvmStatic
-        val postgres = PostgreSQLContainer("postgres:16")
-            .withDatabaseName("testdb")
-            .withUsername("test")
-            .withPassword("test")
-
-        @DynamicPropertySource
-        @JvmStatic
-        fun configureProperties(registry: DynamicPropertyRegistry) {
-            registry.add("spring.datasource.url", postgres::getJdbcUrl)
-            registry.add("spring.datasource.username", postgres::getUsername)
-            registry.add("spring.datasource.password", postgres::getPassword)
-        }
-    }
+) : BaseTestContainer() {
 
     @Test
     fun `Worry를 저장하고 조회할 수 있다`() {
